+++ phoenix/monitor/__init__.py
@@ -10,9 +10,6 @@ def includeme(config):
     config.add_route('monitor', '/monitor')
     config.add_route('monitor_details', '/monitor/details/{job_id}/{tab}')
 
-    # actions
-    config.add_route('update_myjobs', '/monitor/update.json')
-
     config.include('phoenix.monitor.views.actions')