+++ phoenix/static/scripts/phoenix.js
@@ -8,6 +8,57 @@ $(function() {
     $('input[type=checkbox]').prop('checked', this.checked);
   });
 
+  // Open job caption form when caption is clicked
+  $(".caption").button({
+    text: false,
+  }).click(function( event ) {
+    var job_id = $(this).attr('data-value');
+    $.getJSON(
+      '/edit_job.json',
+      {'job_id': job_id},
+      function(json) {
+        if (json) {
+          form = $('#caption-form');
+          
+          // Set the title
+          //form.find('h4').text('Edit Caption');
+          $.each(json, function(k, v) {
+            // Set the value for each field from the returned json
+            form.find('input[name="' + k + '"]').attr('value', v);
+          });
+          
+          form.modal('show');
+        }
+      }
+    );
+  });
+
+
+  // Open job caption form when caption is clicked
+  $(".labels").button({
+    text: false,
+  }).click(function( event ) {
+    var job_id = $(this).attr('data-value');
+    $.getJSON(
+      '/edit_job.json',
+      {'job_id': job_id},
+      function(json) {
+        if (json) {
+          form = $('#labels-form');
+          
+          // Set the title
+          //form.find('h4').text('Edit labels');
+          $.each(json, function(k, v) {
+            // Set the value for each field from the returned json
+            form.find('input[name="' + k + '"]').attr('value', v);
+          });
+          
+          form.modal('show');
+        }
+      }
+    );
+  });
+
   
   // Open publish form when publish is clicked
   $(".publish").button({
@@ -58,5 +109,29 @@ $(function() {
       }
     );
   });
+
+  var selectStatus = function() {
+    default_location = '/monitor/';
+    var location = $('#status-select option:selected').val();
+    if (location){
+      window.location = location;
+    } else {
+      window.location = default_location;
+    }
+  };
+
+  $('#status-select').change( selectStatus );
+
+  var selectLimit = function() {
+    default_location = '/monitor/';
+    var location = $('#limit-select option:selected').val();
+    if (location){
+      window.location = location;
+    } else {
+      window.location = default_location;
+    }
+  };
+
+  $('#limit-select').change( selectLimit );
 
 });