+++ phoenix/tests/test_settings.py
@@ -1,5 +1,5 @@
+import pytest
 import unittest
-from nose import SkipTest
 
 from pyramid import testing
 
@@ -10,8 +10,8 @@ class UserSettingsTests(unittest.TestCase):
     def tearDown(self):
         testing.tearDown()
 
+    @pytest.mark.skip(reason="no way of currently testing this")
     def test_user_view(self):
-        raise SkipTest
         from phoenix.views import UserSettings
 
         request = testing.DummyRequest()
@@ -28,8 +28,8 @@ class UserSettingsFunctionalTests(unittest.TestCase):
 
         #self.testapp = TestApp(app)
 
+    @pytest.mark.skip(reason="no way of currently testing this")
     def test_user_view(self):
-        raise SkipTest
         res = self.testapp.get('/', status=200)
         self.assertIn(b'<h1>Hi Home View', res.body)