+++ phoenix/wizard/views/literalinputs.py
@@ -4,6 +4,8 @@ from owslib.wps import WebProcessingService
 
 from phoenix.wizard.views import Wizard
 from phoenix.wps import WPSSchema
+from phoenix.utils import wps_describe_url
+
 from twitcher.registry import proxy_url
 
 import logging
@@ -12,9 +14,9 @@ logger = logging.getLogger(__name__)
 class LiteralInputs(Wizard):
     def __init__(self, request):
         super(LiteralInputs, self).__init__(request, name='wizard_literal_inputs', title="Literal Inputs")
-        wps = WebProcessingService(url=proxy_url(request, self.wizard_state.get('wizard_wps')['identifier']),
+        self.wps = WebProcessingService(url=proxy_url(request, self.wizard_state.get('wizard_wps')['identifier']),
                                     verify=False, skip_caps=True)
-        self.process = wps.describeprocess(self.wizard_state.get('wizard_process')['identifier'])
+        self.process = self.wps.describeprocess(self.wizard_state.get('wizard_process')['identifier'])
         self.title = "Literal inputs of {0}".format(self.process.title)
 
     def breadcrumbs(self):
@@ -29,7 +31,14 @@ class LiteralInputs(Wizard):
         self.success(appstruct)
         return self.next('wizard_complex_inputs')
     
-    @view_config(route_name='wizard_literal_inputs', renderer='../templates/wizard/default.pt')
+    @view_config(route_name='wizard_literal_inputs', renderer='../templates/wizard/inputs.pt')
     def view(self):
         return super(LiteralInputs, self).view()
-    
+
+    def custom_view(self):
+        return dict(
+            summary_title=self.process.title,
+            summary=getattr(self.process, 'abstract', 'No summary'),
+            url=wps_describe_url(self.wps.url, self.process.identifier),
+            metadata=self.process.metadata)
+