+++ phoenix/wizard/views/source.py
@@ -1,23 +1,29 @@
+import colander
+from deform.widget import RadioChoiceWidget
+
 from pyramid.view import view_config
+
+from owslib.wps import WebProcessingService
+from twitcher.registry import proxy_url
 
 from phoenix.wizard.views import Wizard
 
-import colander
-from deform.widget import RadioChoiceWidget
+SOURCE_TYPES = {
+    'wizard_esgf_search': "Earth System Grid (ESGF)",
+    #'wizard_swift_login': "Swift Cloud",
+    'wizard_threddsservice': "Thredds Catalog Service",
+    'wizard_upload': "Local Storage",
+    'wizard_solr': "Birdhouse Solr Search"
+    }
 
 class SourceSchemaNode(colander.SchemaNode):
     schema_type = colander.String
 
     def after_bind(self, node, kw):
-        choices = [
-            ('wizard_esgf_search', "Earth System Grid (ESGF)"),
-            #('wizard_swift_login', "Swift Cloud"),
-            ('wizard_threddsservice', "Thredds Catalog Service"),
-            ('wizard_upload', "Local Storage"),           
-            ]
-        if kw['request'].solr_activated:
-            choices.append( ('wizard_solr', "Birdhouse Solr Search") )
-        self.widget = RadioChoiceWidget(values = choices)
+        values = SOURCE_TYPES.items()
+        if not kw['request'].solr_activated:
+            values.remove(('wizard_solr', SOURCE_TYPES['wizard_solr']))
+        self.widget = RadioChoiceWidget(values=values)
 
 class Schema(colander.MappingSchema):
     source = SourceSchemaNode()
@@ -26,7 +32,14 @@ class ChooseSource(Wizard):
     def __init__(self, request):
         super(ChooseSource, self).__init__(
             request, name='wizard_source', title="Choose Data Source")
-        self.description = self.wizard_state.get('wizard_complex_inputs')['identifier']
+        wps = WebProcessingService(proxy_url(request, self.wizard_state.get('wizard_wps')['identifier']),
+                                        verify=False, skip_caps=True)
+        process = wps.describeprocess(self.wizard_state.get('wizard_process')['identifier'])
+        for data_input in process.dataInputs:
+            if data_input.identifier == self.wizard_state.get('wizard_complex_inputs')['identifier']:
+                self.title = "Choose Data Source for %s" % data_input.title
+                break
+        #self.description = self.wizard_state.get('wizard_complex_inputs')['identifier']
 
     def breadcrumbs(self):
         breadcrumbs = super(ChooseSource, self).breadcrumbs()